pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]  # Let's use hatchling for now - more stable
build-backend = "hatchling.build"

[project]
name = "deepseek-assistant"
version = "0.1.0"
description = "Local-first AI pair programming assistant with DeepSeek integration."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10,<3.13"
authors = [
    { name = "Stepan Oskin", email = "stepanoskin007@gmail.com" }
]
keywords = [
    "AI", "LLM", "DeepSeek", "code", "assistant", "RAG", "developer-tools",
    "cli", "pair-programming", "local-first"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Environment :: Console",
    "Topic :: Software Development",
    "Topic :: Utilities",
]

# Core runtime dependencies for the assistant package
# HTTP client, YAML config, async file handling, env management, rich CLI output, CLI framework
dependencies = [
    "httpx>=0.24",
    "PyYAML>=6.0",
    "aiofiles>=23.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "click>=8.1",
    "typing-extensions>=4.0",
    "anyio>=4.0",
]

[project.optional-dependencies]
# Development tooling: tests, format, lint, type check, coverage, docs
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "coverage[toml]>=7.4",
    "black>=24.0",
    "ruff>=0.5.0",
    "mypy>=1.8",
    "types-PyYAML>=6.0.12",
    "types-requests>=2.31.0.20240218",
    "pre-commit>=3.0",
]

analysis = [
    "tree-sitter>=0.21.0",
    "tree-sitter-languages>=1.8.0",
    "libcst>=1.1",
    "asttokens>=2.4",
    "parso>=0.8.3",
]

rag = [
    "sentence-transformers>=2.3",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    # Use chromadb as a cross-platform alternative to faiss
    "chromadb>=0.4.0",
]

git = [
    "GitPython>=3.1",
]

web = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
]

# A convenient bundle of all optional features (excluding dev tools)
full = [
    # Include via reference to other groups
    "deepseek-assistant[analysis]",
    "deepseek-assistant[rag]",
    "deepseek-assistant[git]",
    "deepseek-assistant[web]",
]

[project.scripts]
# Primary CLI entry point
deepseek = "assistant.cli.main:cli"
# Alternative/legacy names for convenience
ds = "assistant.cli.main:cli"
deepseek-chat = "assistant.cli.main:cli"

[project.urls]
Homepage = "https://github.com/stpOSkin/deepseek-code-assistant"
Repository = "https://github.com/stpOSkin/deepseek-code-assistant"
"Issue Tracker" = "https://github.com/stpOSkin/deepseek-code-assistant/issues"
Documentation = "https://github.com/stepan-o/deepseek-code-assistant"

# Hatchling configuration for multi-package layout
[tool.hatch.build]
skip-excluded-dirs = true

[tool.hatch.build.targets.wheel]
packages = [
    "src/assistant",
    "src/snapshotter",
]

# uv configuration (just for dependency management)
[tool.uv]
managed = true
package = true  # This just says "uv manages this package" - should be boolean

# Development tool configurations
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.(py|pyi)$'
exclude = [
    "/build/",
    "/dist/",
    "/.venv/",
    "/.git/",
    "/.mypy_cache/",
    "/.ruff_cache/",
    "/__pycache__/",
]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["build", "dist", ".venv"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort (import sorting)
    "B",  # bugbear
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
known_first_party = ["src"]
known_third_party = ["click", "rich", "httpx", "yaml", "dotenv"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
no_implicit_optional = true
strict_equality = true
pretty = true
exclude = [
    "build/",
    "dist/",
    ".venv/",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-v --tb=short"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["src"]
omit = [
    "**/tests/**",
    "**/__pycache__/**",
    ".venv/**",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
